[Feat] #48 - 카페 상세 정보 api
- 카페 메뉴 조회를 위한 api 및 remoteData, repository 작성
- 필요없는 데모데이터 삭제
- item_cafe_info_menu.xml에 databinding 적용

+++ app/src/main/java/org/cazait/data/model/CafeReview.kt
@@ -3,8 +3,8 @@ package org.cazait.data.model
 import com.google.gson.annotations.SerializedName
 
 data class CafeReview(
-    @SerializedName("userId") val userId: Int,
-    @SerializedName("cafeId") val cafeId: Int,
+    @SerializedName("userId") val userId: Long,
+    @SerializedName("cafeId") val cafeId: Long,
     @SerializedName("score") val score: Int,
     @SerializedName("content") val content: String
 )
